feat(security): preserve shutdown codes and cancellation responses
## What

- Preserve non-zero DI shutdown exit codes, including negative values, in `Application.RunCode`.
- Add `StatusClientClosedRequest` for HTTP 499 and render canceled gRPC statuses as `http: client closed request`.
- Reject empty direct PASETO issuer configuration at generate and verify time.
- Document read-only byte ownership for zero-allocation token/source byte APIs and add reviewer guidance for intentional process-global startup behavior.
- Split CLI server/client tests and move shared CLI test setup into `internal/test`.

## Why

- These changes close edge cases found during the code-issue pass while preserving the repository's supported DI/runtime contracts.
- The regression coverage protects CLI shutdown semantics, PASETO issuer validation, and HTTP rendering for client-canceled gRPC requests.
- The test split and reviewer guidance reduce future false positives and make the affected test coverage easier to navigate.

## Testing

- `go test ./os ./token/paseto ./token ./cli ./internal/test ./net/http/status ./net/http/content` - passed
- `git diff --check` - passed
- `make lint` - passed
- `make specs` - passed
